# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Reimbursement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up as a new admin (creates company automatically)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Successfully signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithCompany"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /approval-rules:
    get:
      operationId: listApprovalRules
      tags: [approval-rules]
      summary: List all approval rules (Admin only)
      responses:
        "200":
          description: List of approval rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalRule"
    post:
      operationId: createApprovalRule
      tags: [approval-rules]
      summary: Create an approval rule (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRuleRequest"
      responses:
        "201":
          description: Created approval rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRule"
  /approval-rules/{ruleId}:
    get:
      operationId: getApprovalRule
      tags: [approval-rules]
      summary: Get an approval rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approval rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRule"
    put:
      operationId: updateApprovalRule
      tags: [approval-rules]
      summary: Update an approval rule (Admin only)
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRuleRequest"
      responses:
        "200":
          description: Updated approval rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRule"
    delete:
      operationId: deleteApprovalRule
      tags: [approval-rules]
      summary: Delete an approval rule (Admin only)
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        country:
          type: string
        currency:
          type: string
      required: [email, password, firstName, lastName, companyName, country, currency]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserWithCompany"
        token:
          type: string
      required: [user, token]
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        currency:
          type: string
        currencySymbol:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, country, currency, currencySymbol, createdAt]
...
    UserWithCompany:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        isManagerApprover:
          type: boolean
        managerId:
          type: integer
          nullable: true
        companyId:
          type: integer
        isActive:
          type: boolean
        company:
          $ref: "#/components/schemas/Company"
        createdAt:
          type: string
          format: date-time
      required: [id, email, firstName, lastName, role, isManagerApprover, companyId, isActive, company, createdAt]
...
    ApprovalRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        companyId:
          type: integer
        isDefault:
          type: boolean
        conditionType:
          type: string
          enum: [sequential, percentage, specific_approver, hybrid]
        percentageThreshold:
          type: number
          nullable: true
        specificApproverId:
          type: integer
          nullable: true
        specificApproverName:
          type: string
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalStep"
        createdAt:
          type: string
          format: date-time
      required: [id, name, companyId, isDefault, conditionType, steps, createdAt]
    ApprovalStep:
      type: object
      properties:
        id:
          type: integer
        ruleId:
          type: integer
        stepNumber:
          type: integer
        approverId:
          type: integer
        approverName:
          type: string
        approverRole:
          type: string
        isManagerStep:
          type: boolean
      required: [id, ruleId, stepNumber, approverId, approverName, approverRole, isManagerStep]
    CreateApprovalRuleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        isDefault:
          type: boolean
        conditionType:
          type: string
          enum: [sequential, percentage, specific_approver, hybrid]
        percentageThreshold:
          type: number
          nullable: true
        specificApproverId:
          type: integer
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/CreateApprovalStepRequest"
      required: [name, conditionType, steps]
    CreateApprovalStepRequest:
      type: object
      properties:
        stepNumber:
          type: integer
        approverId:
          type: integer
        isManagerStep:
          type: boolean
      required: [stepNumber, approverId, isManagerStep]
...