# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reimbursement Management System API
 * OpenAPI spec version: 1.0.0
 */

export interface Company {
  id: number;
  name: string;
  country: string;
  currency: string;
  currencySymbol: string;
  createdAt: Date;
}